+++ src/vga/mod.rs
@@ -12,7 +12,7 @@
 // -----------------------------------------------------------------------------
 // Licence Statement
 // -----------------------------------------------------------------------------
-// Copyright (c) Jonathan 'theJPster' Pallant and the Neotron Developers, 2021
+// Copyright (c) Jonathan 'theJPster' Pallant and the Neotron Developers, 2023
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the Free Software
@@ -34,19 +34,79 @@
 
 mod font16;
 mod font8;
+mod rgb;
 
 // -----------------------------------------------------------------------------
 // Imports
 // -----------------------------------------------------------------------------
 
-use core::sync::atomic::{AtomicBool, AtomicPtr, AtomicU16, AtomicU8, Ordering};
+use core::sync::atomic::{AtomicBool, AtomicU16, AtomicU8, Ordering};
 use defmt::{debug, trace};
-use neotron_common_bios::video::{Attr, TextBackgroundColour, TextForegroundColour};
+use neotron_common_bios::video::{Attr, GlyphAttr, TextBackgroundColour, TextForegroundColour};
 use rp_pico::{hal::pio::PIOExt, pac::interrupt};
+
+pub use rgb::{RGBColour, RGBPair};
 
 // -----------------------------------------------------------------------------
 // Types
 // -----------------------------------------------------------------------------
+
+/// A font
+pub struct Font<'a> {
+	data: &'a [u8],
+}
+
+/// Holds a video mode, but as an atomic value suitable for use in a `static`.
+pub struct AtomicModeWrapper {
+	value: AtomicU8,
+}
+
+impl AtomicModeWrapper {
+	/// Construct a new [`AtomicModeWrapper`] with the given value.
+	const fn new(mode: crate::common::video::Mode) -> AtomicModeWrapper {
+		AtomicModeWrapper {
+			value: AtomicU8::new(mode.as_u8()),
+		}
+	}
+
+	/// Set a new video mode.
+	pub fn set_mode(&self, mode: crate::common::video::Mode) {
+		self.value.store(mode.as_u8(), Ordering::SeqCst);
+	}
+
+	/// Get the current video mode.
+	pub fn get_mode(&self) -> crate::common::video::Mode {
+		let value = self.value.load(Ordering::SeqCst);
+		// Safety: the 'set_mode' function ensure this is always valid.
+		unsafe { crate::common::video::Mode::from_u8(value) }
+	}
+}
+
+/// Describes the polarity of a sync pulse.
+///
+/// Some pulses are positive (active-high), some are negative (active-low).
+enum SyncPolarity {
+	/// An active-high pulse
+	Positive,
+	/// An active-low pulse
+	Negative,
+}
+
+impl SyncPolarity {
+	const fn enabled(&self) -> bool {
+		match self {
+			SyncPolarity::Positive => true,
+			SyncPolarity::Negative => false,
+		}
+	}
+
+	const fn disabled(&self) -> bool {
+		match self {
+			SyncPolarity::Positive => false,
+			SyncPolarity::Negative => true,
+		}
+	}
+}
 
 /// Holds some data necessary to run the Video.
 ///
@@ -63,19 +123,151 @@ struct RenderEngine {
 	num_text_cols: usize,
 }
 
-/// A font
-pub struct Font<'a> {
-	data: &'a [u8],
+impl RenderEngine {
+	// Initialise the main-thread resources
+	pub fn new() -> RenderEngine {
+		RenderEngine {
+			frame_count: 0,
+			// Should match the default value of TIMING_BUFFER and VIDEO_MODE
+			current_video_mode: crate::common::video::Mode::new(
+				crate::common::video::Timing::T640x480,
+				crate::common::video::Format::Text8x16,
+			),
+			// Should match the mode above
+			num_text_cols: 80,
+			num_text_rows: 30,
+		}
+	}
+
+	/// Do the start-of-frame setup
+	fn frame_start(&mut self) {
+		trace!("Frame {}", self.frame_count);
+		self.frame_count += 1;
+
+		// Update video mode only on first line of video
+		if VIDEO_MODE.get_mode() != self.current_video_mode {
+			self.current_video_mode = VIDEO_MODE.get_mode();
+			match self.current_video_mode.timing() {
+				crate::common::video::Timing::T640x400 => unsafe {
+					TIMING_BUFFER = TimingBuffer::make_640x400();
+				},
+				crate::common::video::Timing::T640x480 => unsafe {
+					TIMING_BUFFER = TimingBuffer::make_640x480();
+				},
+				crate::common::video::Timing::T800x600 => {
+					panic!("Can't do 800x600");
+				}
+			}
+			self.num_text_cols = self.current_video_mode.text_width().unwrap_or(0) as usize;
+			self.num_text_rows = self.current_video_mode.text_height().unwrap_or(0) as usize;
+		}
+	}
+
+	/// Draw a line of pixels into the relevant pixel buffer (either
+	/// [`PIXEL_DATA_BUFFER_ODD`] or [`PIXEL_DATA_BUFFER_EVEN`]).
+	///
+	/// The `current_line_num` goes from `0..NUM_LINES`.
+	#[link_section = ".data"]
+	pub fn draw_next_line(&mut self, current_line_num: u16) {
+		// Pick a buffer to render into based on the line number we are drawing.
+		// It's safe to write to this buffer because it's the the other one that
+		// is currently being DMA'd out to the Pixel SM.
+		let scan_line_buffer = unsafe {
+			if (current_line_num & 1) == 0 {
+				&mut PIXEL_DATA_BUFFER_EVEN
+			} else {
+				&mut PIXEL_DATA_BUFFER_ODD
+			}
+		};
+
+		match self.current_video_mode.format() {
+			crate::common::video::Format::Text8x16 => {
+				self.draw_next_line_text::<16>(&font16::FONT, scan_line_buffer, current_line_num)
+			}
+			crate::common::video::Format::Text8x8 => {
+				self.draw_next_line_text::<8>(&font8::FONT, scan_line_buffer, current_line_num)
+			}
+			_ => {
+				// Draw nothing
+			}
+		};
+	}
+
+	/// Draw a line of pixels into the relevant pixel buffer (either
+	/// [`PIXEL_DATA_BUFFER_ODD`] or [`PIXEL_DATA_BUFFER_EVEN`]) using the 8x16
+	/// font.
+	///
+	/// The `current_line_num` goes from `0..NUM_LINES`.
+	#[link_section = ".data"]
+	pub fn draw_next_line_text<const GLYPH_HEIGHT: usize>(
+		&mut self,
+		font: &Font,
+		scan_line_buffer: &mut LineBuffer,
+		current_line_num: u16,
+	) {
+		// Convert our position in scan-lines to a text row, and a line within each glyph on that row
+		let text_row = current_line_num as usize / GLYPH_HEIGHT;
+		let font_row = current_line_num as usize % GLYPH_HEIGHT;
+
+		if text_row < self.num_text_rows {
+			// Note (unsafe): We could stash the char array inside `self`
+			// but at some point we are going to need one CPU rendering
+			// the text, and the other CPU running code and writing to
+			// the buffer. This might be Undefined Behaviour, but
+			// unfortunately real-time video is all about shared mutable
+			// state. At least our platform is fixed, so we can simply
+			// test if it works, for some given version of the Rust compiler.
+			let row_slice = unsafe {
+				&GLYPH_ATTR_ARRAY
+					[(text_row * self.num_text_cols)..((text_row + 1) * self.num_text_cols)]
+			};
+			// Every font look-up we are about to do for this row will
+			// involve offsetting by the row within each glyph. As this
+			// is the same for every glyph on this row, we calculate a
+			// new pointer once, in advance, and save ourselves an
+			// addition each time around the loop.
+			let font_ptr = unsafe { font.data.as_ptr().add(font_row) };
+
+			// Get a pointer into our scan-line buffer
+			let mut scan_line_buffer_ptr = scan_line_buffer.pixels.as_mut_ptr();
+
+			// Convert from characters to coloured pixels, using the font as a look-up table.
+
+			for glyphattr in row_slice.iter() {
+				let index = (glyphattr.glyph().0 as usize) * GLYPH_HEIGHT;
+				let attr = glyphattr.attr();
+
+				unsafe {
+					// Note (unsafe): We use pointer arithmetic here because we
+					// can't afford a bounds-check on an array. This is safe
+					// because the font is `256 * width` bytes long and we can't
+					// index more than `255 * width` bytes into it.
+					let mono_pixels = *font_ptr.add(index);
+
+					// 0bXX------
+					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels >> 6);
+					scan_line_buffer_ptr.write(pair);
+					// 0b--XX----
+					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels >> 4);
+					scan_line_buffer_ptr.offset(1).write(pair);
+					// 0b----XX--
+					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels >> 2);
+					scan_line_buffer_ptr.offset(2).write(pair);
+					// 0b------XX
+					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels);
+					scan_line_buffer_ptr.offset(3).write(pair);
+
+					scan_line_buffer_ptr = scan_line_buffer_ptr.offset(4);
+				}
+			}
+		}
+	}
 }
 
-/// Holds some data necessary to present a text console.
-///
-/// Used by Core 0 to control writes to a shared text-buffer.
-pub struct TextConsole {
-	current_col: AtomicU16,
-	current_row: AtomicU16,
-	text_buffer: AtomicPtr<GlyphAttr>,
-	current_attr: AtomicU8,
+impl Default for RenderEngine {
+	fn default() -> Self {
+		RenderEngine::new()
+	}
 }
 
 /// Describes one scan-line's worth of pixels, including the length word required by the Pixel FIFO.
@@ -87,14 +279,11 @@ struct LineBuffer {
 	pixels: [RGBPair; MAX_NUM_PIXEL_PAIRS_PER_LINE],
 }
 
-/// Describes the polarity of a sync pulse.
-///
-/// Some pulses are positive (active-high), some are negative (active-low).
-pub enum SyncPolarity {
-	/// An active-high pulse
-	Positive,
-	/// An active-low pulse
-	Negative,
+impl LineBuffer {
+	/// Convert the line buffer to a 32-bit address that the DMA engine understands.
+	fn as_ptr(&self) -> u32 {
+		self as *const _ as usize as u32
+	}
 }
 
 /// Holds the four scan-line timing FIFO words we need for one scan-line.
@@ -104,6 +293,173 @@ pub enum SyncPolarity {
 #[repr(C, align(16))]
 struct ScanlineTimingBuffer {
 	data: [u32; 4],
+}
+
+impl ScanlineTimingBuffer {
+	const CLOCKS_PER_PIXEL: u32 = 6;
+
+	/// Create a timing buffer for each scan-line in the V-Sync visible portion.
+	///
+	/// The timings are in the order (front-porch, sync, back-porch, visible) and are in pixel clocks.
+	const fn new_v_visible(
+		hsync: SyncPolarity,
+		vsync: SyncPolarity,
+		timings: (u32, u32, u32, u32),
+	) -> ScanlineTimingBuffer {
+		ScanlineTimingBuffer {
+			data: [
+				// Front porch (as per the spec)
+				Self::make_timing(
+					timings.0 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.disabled(),
+					Some(1),
+				),
+				// Sync pulse (as per the spec)
+				Self::make_timing(
+					timings.1 * Self::CLOCKS_PER_PIXEL,
+					hsync.enabled(),
+					vsync.disabled(),
+					None,
+				),
+				// Back porch. Adjusted by a few clocks to account for interrupt +
+				// PIO SM start latency.
+				Self::make_timing(
+					(timings.2 * Self::CLOCKS_PER_PIXEL) - 5,
+					hsync.disabled(),
+					vsync.disabled(),
+					None,
+				),
+				// Visible portion. It also triggers the IRQ to start pixels
+				// moving. Adjusted to compensate for changes made to previous
+				// period to ensure scan-line remains at correct length.
+				Self::make_timing(
+					(timings.3 * Self::CLOCKS_PER_PIXEL) + 5,
+					hsync.disabled(),
+					vsync.disabled(),
+					Some(0),
+				),
+			],
+		}
+	}
+
+	/// Create a timing buffer for each scan-line in the V-Sync front-porch and back-porch
+	const fn new_v_porch(
+		hsync: SyncPolarity,
+		vsync: SyncPolarity,
+		timings: (u32, u32, u32, u32),
+	) -> ScanlineTimingBuffer {
+		ScanlineTimingBuffer {
+			data: [
+				// Front porch (as per the spec)
+				Self::make_timing(
+					timings.0 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.disabled(),
+					Some(1),
+				),
+				// Sync pulse (as per the spec)
+				Self::make_timing(
+					timings.1 * Self::CLOCKS_PER_PIXEL,
+					hsync.enabled(),
+					vsync.disabled(),
+					None,
+				),
+				// Back porch.
+				Self::make_timing(
+					timings.2 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.disabled(),
+					None,
+				),
+				// Visible portion.
+				Self::make_timing(
+					timings.3 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.disabled(),
+					None,
+				),
+			],
+		}
+	}
+
+	/// Create a timing buffer for each scan-line in the V-Sync pulse
+	const fn new_v_pulse(
+		hsync: SyncPolarity,
+		vsync: SyncPolarity,
+		timings: (u32, u32, u32, u32),
+	) -> ScanlineTimingBuffer {
+		ScanlineTimingBuffer {
+			data: [
+				// Front porch (as per the spec)
+				Self::make_timing(
+					timings.0 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.enabled(),
+					Some(1),
+				),
+				// Sync pulse (as per the spec)
+				Self::make_timing(
+					timings.1 * Self::CLOCKS_PER_PIXEL,
+					hsync.enabled(),
+					vsync.enabled(),
+					None,
+				),
+				// Back porch.
+				Self::make_timing(
+					timings.2 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.enabled(),
+					None,
+				),
+				// Visible portion.
+				Self::make_timing(
+					timings.3 * Self::CLOCKS_PER_PIXEL,
+					hsync.disabled(),
+					vsync.enabled(),
+					None,
+				),
+			],
+		}
+	}
+
+	/// Generate a 32-bit value we can send to the Timing FIFO.
+	///
+	/// * `period` - The length of this portion of the scan-line, in system clock ticks
+	/// * `hsync` - true if the H-Sync pin should be high during this period, else false
+	/// * `vsync` - true if the H-Sync pin should be high during this period, else false
+	/// * `raise_irq` - true the timing statemachine should raise an IRQ at the start of this period
+	///
+	/// Returns a 32-bit value you can post to the Timing FIFO.
+	const fn make_timing(period: u32, hsync: bool, vsync: bool, raise_irq: Option<u8>) -> u32 {
+		let command = if let Some(irq_index) = raise_irq {
+			// This command sets IRQ 0
+			pio::InstructionOperands::IRQ {
+				clear: false,
+				wait: false,
+				index: irq_index,
+				relative: false,
+			}
+			.encode()
+		} else {
+			// This command is a no-op (it moves Y into Y)
+			pio::InstructionOperands::MOV {
+				destination: pio::MovDestination::Y,
+				op: pio::MovOperation::None,
+				source: pio::MovSource::Y,
+			}
+			.encode()
+		} as u32;
+		let mut value: u32 = 0;
+		if hsync {
+			value |= 1 << 0;
+		}
+		if vsync {
+			value |= 1 << 1;
+		}
+		value |= (period - 6) << 2;
+		value | command << 16
+	}
 }
 
 /// Holds the different kinds of scan-line timing buffers we need for various
@@ -125,42 +481,105 @@ struct TimingBuffer {
 	back_porch_ends_at: u16,
 }
 
-/// Represents a 12-bit colour value.
-///
-/// Each channel has four-bits, and they are packed in `BGR` format. This is
-/// so the PIO can shift them out right-first, and we have RED0 assigned to
-/// the lowest GPIO pin.
-#[repr(transparent)]
-#[derive(Copy, Clone, PartialEq, Eq)]
-pub struct RGBColour(pub u16);
+impl TimingBuffer {
+	/// Make a timing buffer suitable for 640 x 400 @ 70 Hz
+	const fn make_640x400() -> TimingBuffer {
+		TimingBuffer {
+			visible_line: ScanlineTimingBuffer::new_v_visible(
+				SyncPolarity::Negative,
+				SyncPolarity::Positive,
+				(16, 96, 48, 640),
+			),
+			vblank_porch_buffer: ScanlineTimingBuffer::new_v_porch(
+				SyncPolarity::Negative,
+				SyncPolarity::Positive,
+				(16, 96, 48, 640),
+			),
+			vblank_sync_buffer: ScanlineTimingBuffer::new_v_pulse(
+				SyncPolarity::Negative,
+				SyncPolarity::Positive,
+				(16, 96, 48, 640),
+			),
+			visible_lines_ends_at: 399,
+			front_porch_end_at: 399 + 12,
+			sync_pulse_ends_at: 399 + 12 + 2,
+			back_porch_ends_at: 399 + 12 + 2 + 35,
+		}
+	}
 
-/// Represents two `RGBColour` pixels packed together.
-///
-/// The `first` pixel is packed in the lower 16-bits. This is because the PIO
-/// shifts-right.
-#[repr(transparent)]
-#[derive(Copy, Clone, PartialEq, Eq)]
-pub struct RGBPair(u32);
-
-/// Represents a glyph in the current font.
-#[repr(transparent)]
-#[derive(Copy, Clone, PartialEq, Eq)]
-pub struct Glyph(pub u8);
-
-/// Represents a glyph/attribute pair. This is what out text console is made
-/// out of. They work in exactly the same way as IBM PC VGA.
-#[repr(transparent)]
-#[derive(Copy, Clone, PartialEq, Eq, Default)]
-pub struct GlyphAttr(u16);
-
-/// Holds a video mode, but as an atomic value suitable for use in a `static`.
-struct AtomicModeWrapper {
-	value: AtomicU8,
+	/// Make a timing buffer suitable for 640 x 480 @ 60 Hz
+	pub const fn make_640x480() -> TimingBuffer {
+		TimingBuffer {
+			visible_line: ScanlineTimingBuffer::new_v_visible(
+				SyncPolarity::Negative,
+				SyncPolarity::Negative,
+				(16, 96, 48, 640),
+			),
+			vblank_porch_buffer: ScanlineTimingBuffer::new_v_porch(
+				SyncPolarity::Negative,
+				SyncPolarity::Negative,
+				(16, 96, 48, 640),
+			),
+			vblank_sync_buffer: ScanlineTimingBuffer::new_v_pulse(
+				SyncPolarity::Negative,
+				SyncPolarity::Negative,
+				(16, 96, 48, 640),
+			),
+			visible_lines_ends_at: 479,
+			front_porch_end_at: 479 + 10,
+			sync_pulse_ends_at: 479 + 10 + 2,
+			back_porch_ends_at: 479 + 10 + 2 + 33,
+		}
+	}
 }
 
 /// See [`TEXT_COLOUR_LOOKUP`]
 struct TextColourLookup {
 	entries: [RGBPair; 512],
+}
+
+impl TextColourLookup {
+	const fn blank() -> TextColourLookup {
+		TextColourLookup {
+			entries: [RGBPair::new(); 512],
+		}
+	}
+
+	fn init(&mut self, palette: &[AtomicU16]) {
+		for (fg, fg_colour) in palette.iter().take(16).enumerate() {
+			for (bg, bg_colour) in palette.iter().take(8).enumerate() {
+				let attr = unsafe {
+					Attr::new(
+						TextForegroundColour::new_unchecked(fg as u8),
+						TextBackgroundColour::new_unchecked(bg as u8),
+						false,
+					)
+				};
+				for pixels in 0..=3 {
+					let index: usize = (((attr.0 & 0x7F) as usize) << 2) | (pixels & 0x03) as usize;
+					let pair = RGBPair::from_pixels(
+						if pixels & 0x02 == 0x02 {
+							RGBColour(fg_colour.load(Ordering::Relaxed))
+						} else {
+							RGBColour(bg_colour.load(Ordering::Relaxed))
+						},
+						if pixels & 0x01 == 0x01 {
+							RGBColour(fg_colour.load(Ordering::Relaxed))
+						} else {
+							RGBColour(bg_colour.load(Ordering::Relaxed))
+						},
+					);
+					self.entries[index] = pair;
+				}
+			}
+		}
+	}
+
+	#[inline]
+	fn lookup(&self, attr: Attr, pixels: u8) -> RGBPair {
+		let index: usize = (((attr.0 & 0x7F) as usize) << 2) | (pixels & 0x03) as usize;
+		unsafe { core::ptr::read(self.entries.as_ptr().add(index)) }
+	}
 }
 
 // -----------------------------------------------------------------------------
@@ -172,77 +591,16 @@ struct TextColourLookup {
 /// Adjust the pixel PIO program to run at the right speed to the screen is
 /// filled. For example, if this is only 320 but you are aiming at 640x480,
 /// make the pixel PIO take twice as long per pixel.
-const MAX_NUM_PIXELS_PER_LINE: usize = 640;
+pub const MAX_NUM_PIXELS_PER_LINE: usize = 640;
 
 /// Maximum number of lines on screen.
-const MAX_NUM_LINES: usize = 480;
-
-/// How many pixel pairs we send out.
-///
-/// Each pixel is two 12-bit values packed into one 32-bit word(an `RGBPair`).
-/// This is to make more efficient use of DMA and FIFO resources.
-const MAX_NUM_PIXEL_PAIRS_PER_LINE: usize = MAX_NUM_PIXELS_PER_LINE / 2;
+pub const MAX_NUM_LINES: usize = 480;
 
 /// The highest number of columns in any text mode.
 pub const MAX_TEXT_COLS: usize = MAX_NUM_PIXELS_PER_LINE / 8;
 
 /// The highest number of rows in any text mode.
 pub const MAX_TEXT_ROWS: usize = MAX_NUM_LINES / 8;
-
-/// Used to signal when Core 1 has started
-static CORE1_START_FLAG: AtomicBool = AtomicBool::new(false);
-
-/// Stores our timing data which we DMA into the timing PIO State Machine.
-///
-/// Default value must match [`RenderEngine::new`]
-static mut TIMING_BUFFER: TimingBuffer = TimingBuffer::make_640x480();
-
-/// Stores our current video mode, or the mode we change into on the next frame.
-///
-/// We boot in 80x30 mode.
-static VIDEO_MODE: AtomicModeWrapper = AtomicModeWrapper::new(crate::common::video::Mode::new(
-	crate::common::video::Timing::T640x480,
-	crate::common::video::Format::Text8x16,
-));
-
-/// Tracks which scan-line will be shown next.
-///
-/// This is for timing purposes, therefore it goes from
-/// `0..TIMING_BUFFER.back_porch_ends_at`.
-///
-/// Set by the PIO IRQ.
-static NEXT_SCAN_LINE: AtomicU16 = AtomicU16::new(0);
-
-/// Indicates that we should draw the current scan-line given by [`NEXT_SCAN_LINE`].
-///
-/// Set by the PIO IRQ.
-static DRAW_THIS_LINE: AtomicBool = AtomicBool::new(false);
-
-/// DMA channel for the timing FIFO
-const TIMING_DMA_CHAN: usize = 0;
-
-/// DMA channel for the pixel FIFO
-const PIXEL_DMA_CHAN: usize = 1;
-
-/// One scan-line's worth of 12-bit pixels, used for the even scan-lines (0, 2, 4 ... NUM_LINES-2).
-///
-/// Gets read by DMA, which pushes them into the pixel state machine's FIFO.
-///
-/// Gets written to by `RenderEngine` running on Core 1.
-static mut PIXEL_DATA_BUFFER_EVEN: LineBuffer = LineBuffer {
-	length: (MAX_NUM_PIXEL_PAIRS_PER_LINE as u32) - 1,
-	pixels: [RGBPair::from_pixels(colours::WHITE, colours::BLACK); MAX_NUM_PIXEL_PAIRS_PER_LINE],
-};
-
-/// One scan-line's worth of 12-bit pixels, used for the odd scan-lines (1, 3, 5 ... NUM_LINES-1).
-///
-/// Gets read by DMA, which pushes them into the pixel state machine's FIFO.
-///
-/// Gets written to by `RenderEngine` running on Core 1.
-static mut PIXEL_DATA_BUFFER_ODD: LineBuffer = LineBuffer {
-	length: (MAX_NUM_PIXEL_PAIRS_PER_LINE as u32) - 1,
-	pixels: [RGBPair::from_pixels(colours::BLACK, colours::WHITE); MAX_NUM_PIXEL_PAIRS_PER_LINE],
-};
 
 /// This is our text buffer.
 ///
@@ -253,54 +611,13 @@ static mut PIXEL_DATA_BUFFER_ODD: LineBuffer = LineBuffer {
 pub static mut GLYPH_ATTR_ARRAY: [GlyphAttr; MAX_TEXT_COLS * MAX_TEXT_ROWS] =
 	[GlyphAttr(0); MAX_TEXT_COLS * MAX_TEXT_ROWS];
 
-/// Core 1 entry function.
+/// Stores our current video mode, or the mode we change into on the next frame.
 ///
-/// This is a naked function I have pre-compiled to thumb-2 instructions. I
-/// could use inline assembler, but then I'd have to make you install
-/// arm-none-eabi-as or arm-none-eabi-gcc, or wait until `llvm_asm!` is
-/// stablised.
-///
-/// We want this function to load the three parameters (which we placed in
-/// Core 1's stack) into registers. We then jump to the third of these, which
-/// effectively makes a function call with the first two values as function
-/// arguments.
-static CORE1_ENTRY_FUNCTION: [u16; 2] = [
-	0xbd03, // pop {r0, r1, pc}
-	0x46c0, // nop - pad this out to 32-bits long
-];
-
-/// Holds the colour look-up table for text mode.
-///
-/// The input is a 9-bit value comprised of the 4-bit foreground colour index,
-/// the 3-bit background colour index, and a two mono pixels. The output is a
-/// 32-bit RGB Colour Pair, containing two RGB pixels.
-///
-/// ```
-/// +-----+-----+-----+-----+-----+-----+-----+-----+-----+
-/// | FG3 | FG2 | FG1 | FG0 | BG2 | BG1 | BG0 | PX1 | PX0 |
-/// +-----+-----+-----+-----+-----+-----+-----+-----+-----+
-/// ```
-static mut TEXT_COLOUR_LOOKUP: TextColourLookup = TextColourLookup::blank();
-
-/// A set of useful constants representing common RGB colours.
-pub mod colours {
-	pub const BLACK: super::RGBColour = super::RGBColour::from_24bit(0x00, 0x00, 0x00);
-	pub const DARK_GRAY: super::RGBColour = super::RGBColour::from_24bit(0x80, 0x80, 0x80);
-	pub const BLUE: super::RGBColour = super::RGBColour::from_24bit(0x00, 0x00, 0x80);
-	pub const LIGHT_BLUE: super::RGBColour = super::RGBColour::from_24bit(0x00, 0x00, 0xF0);
-	pub const GREEN: super::RGBColour = super::RGBColour::from_24bit(0x00, 0x80, 0x00);
-	pub const LIGHT_GREEN: super::RGBColour = super::RGBColour::from_24bit(0x00, 0xF0, 0x00);
-	pub const CYAN: super::RGBColour = super::RGBColour::from_24bit(0x00, 0x80, 0x80);
-	pub const LIGHT_CYAN: super::RGBColour = super::RGBColour::from_24bit(0x00, 0xF0, 0xF0);
-	pub const RED: super::RGBColour = super::RGBColour::from_24bit(0x80, 0x00, 0x00);
-	pub const LIGHT_RED: super::RGBColour = super::RGBColour::from_24bit(0xF0, 0x00, 0x00);
-	pub const MAGENTA: super::RGBColour = super::RGBColour::from_24bit(0x80, 0x00, 0x80);
-	pub const LIGHT_MAGENTA: super::RGBColour = super::RGBColour::from_24bit(0xF0, 0x00, 0xF0);
-	pub const BROWN: super::RGBColour = super::RGBColour::from_24bit(0x80, 0x80, 0x00);
-	pub const YELLOW: super::RGBColour = super::RGBColour::from_24bit(0xF0, 0xF0, 0x00);
-	pub const LIGHT_GRAY: super::RGBColour = super::RGBColour::from_24bit(0xA0, 0xA0, 0xA0);
-	pub const WHITE: super::RGBColour = super::RGBColour::from_24bit(0xF0, 0xF0, 0xF0);
-}
+/// We boot in 80x30 mode.
+pub static VIDEO_MODE: AtomicModeWrapper = AtomicModeWrapper::new(crate::common::video::Mode::new(
+	crate::common::video::Timing::T640x480,
+	crate::common::video::Format::Text8x16,
+));
 
 /// Holds the 256-entry palette for indexed colour modes.
 ///
@@ -822,6 +1139,71 @@ pub static VIDEO_PALETTE: [AtomicU16; 256] = [
 	AtomicU16::new(RGBColour::from_12bit(0x7, 0x7, 0x7).0),
 ];
 
+/// How many pixel pairs we send out.
+///
+/// Each pixel is two 12-bit values packed into one 32-bit word(an `RGBPair`).
+/// This is to make more efficient use of DMA and FIFO resources.
+const MAX_NUM_PIXEL_PAIRS_PER_LINE: usize = MAX_NUM_PIXELS_PER_LINE / 2;
+
+/// Stores our timing data which we DMA into the timing PIO State Machine.
+///
+/// Default value must match [`RenderEngine::new`]
+static mut TIMING_BUFFER: TimingBuffer = TimingBuffer::make_640x480();
+
+/// Tracks which scan-line will be shown next.
+///
+/// This is for timing purposes, therefore it goes from
+/// `0..TIMING_BUFFER.back_porch_ends_at`.
+///
+/// Set by the PIO IRQ.
+static NEXT_SCAN_LINE: AtomicU16 = AtomicU16::new(0);
+
+/// Indicates that we should draw the current scan-line given by [`NEXT_SCAN_LINE`].
+///
+/// Set by the PIO IRQ.
+static DRAW_THIS_LINE: AtomicBool = AtomicBool::new(false);
+
+/// DMA channel for the timing FIFO
+const TIMING_DMA_CHAN: usize = 0;
+
+/// DMA channel for the pixel FIFO
+const PIXEL_DMA_CHAN: usize = 1;
+
+/// One scan-line's worth of 12-bit pixels, used for the even scan-lines (0, 2, 4 ... NUM_LINES-2).
+///
+/// Gets read by DMA, which pushes them into the pixel state machine's FIFO.
+///
+/// Gets written to by `RenderEngine` running on Core 1.
+static mut PIXEL_DATA_BUFFER_EVEN: LineBuffer = LineBuffer {
+	length: (MAX_NUM_PIXEL_PAIRS_PER_LINE as u32) - 1,
+	pixels: [RGBPair::from_pixels(RGBColour::WHITE, RGBColour::BLACK);
+		MAX_NUM_PIXEL_PAIRS_PER_LINE],
+};
+
+/// One scan-line's worth of 12-bit pixels, used for the odd scan-lines (1, 3, 5 ... NUM_LINES-1).
+///
+/// Gets read by DMA, which pushes them into the pixel state machine's FIFO.
+///
+/// Gets written to by `RenderEngine` running on Core 1.
+static mut PIXEL_DATA_BUFFER_ODD: LineBuffer = LineBuffer {
+	length: (MAX_NUM_PIXEL_PAIRS_PER_LINE as u32) - 1,
+	pixels: [RGBPair::from_pixels(RGBColour::BLACK, RGBColour::WHITE);
+		MAX_NUM_PIXEL_PAIRS_PER_LINE],
+};
+
+/// Holds the colour look-up table for text mode.
+///
+/// The input is a 9-bit value comprised of the 4-bit foreground colour index,
+/// the 3-bit background colour index, and a two mono pixels. The output is a
+/// 32-bit RGB Colour Pair, containing two RGB pixels.
+///
+/// ```
+/// +-----+-----+-----+-----+-----+-----+-----+-----+-----+
+/// | FG3 | FG2 | FG1 | FG0 | BG2 | BG1 | BG0 | PX1 | PX0 |
+/// +-----+-----+-----+-----+-----+-----+-----+-----+-----+
+/// ```
+static mut TEXT_COLOUR_LOOKUP: TextColourLookup = TextColourLookup::blank();
+
 // -----------------------------------------------------------------------------
 // Functions
 // -----------------------------------------------------------------------------
@@ -1043,7 +1425,7 @@ pub fn init(
 		TEXT_COLOUR_LOOKUP.init(&VIDEO_PALETTE);
 	}
 
-	multicore_launch_core1_with_stack(
+	crate::multicore::launch_core1_with_stack(
 		core1_main,
 		unsafe { &mut super::CORE1_STACK },
 		ppb,
@@ -1052,108 +1434,6 @@ pub fn init(
 	);
 
 	debug!("Core 1 running");
-}
-
-/// The bootrom code will call this function on core1 to perform any set-up, before the
-/// entry function is called.
-extern "C" fn core1_wrapper(entry_func: extern "C" fn() -> u32, _stack_base: *mut u32) -> u32 {
-	entry_func()
-}
-
-/// Starts core 1 running the given function, with the given stack.
-fn multicore_launch_core1_with_stack(
-	main_func: unsafe extern "C" fn() -> u32,
-	stack: &mut [usize],
-	ppb: &mut crate::pac::PPB,
-	fifo: &mut rp_pico::hal::sio::SioFifo,
-	psm: &mut crate::pac::PSM,
-) {
-	debug!("Resetting CPU1...");
-
-	psm.frce_off.modify(|_, w| w.proc1().set_bit());
-	while !psm.frce_off.read().proc1().bit_is_set() {
-		cortex_m::asm::nop();
-	}
-	psm.frce_off.modify(|_, w| w.proc1().clear_bit());
-
-	debug!("Setting up stack...");
-
-	// Gets popped into `r0` by CORE1_ENTRY_FUNCTION. This is the `main`
-	// function we want to run. It appears in the call to `core1_wrapper` as
-	// the first argument.
-	stack[stack.len() - 3] = main_func as *const () as usize;
-	// Gets popped into `r1` by CORE1_ENTRY_FUNCTION. This is the top of stack
-	// for Core 1. It appears in the call to `core1_wrapper` as the second
-	// argument.
-	stack[stack.len() - 2] = stack.as_ptr() as *const _ as usize;
-	// Gets popped into `pc` by CORE1_ENTRY_FUNCTION. This is the function
-	// `CORE1_ENTRY_FUNCTION` will jump to, passing the above two values as
-	// arguments.
-	stack[stack.len() - 1] = core1_wrapper as *const () as usize;
-	// Point into the top of the stack (so there are three values pushed onto
-	// it, i.e. at/above it)
-	let stack_ptr = unsafe { stack.as_mut_ptr().add(stack.len() - 3) };
-
-	debug!("Stack ptr is 0x{:x}", stack_ptr);
-	debug!("Stack bottom is 0x{:x}", stack.as_ptr());
-	debug!("Stack top is 0x{:x}", &stack[stack.len() - 4..stack.len()]);
-
-	// This is the launch sequence we send to core1, to get it to leave the
-	// boot ROM and run our code.
-	let cmd_sequence: [u32; 6] = [
-		0,
-		0,
-		1,
-		ppb.vtor.read().bits() as usize as u32,
-		stack_ptr as usize as u32,
-		// Have to add 1 to convert from an array pointer to a thumb instruction pointer
-		(CORE1_ENTRY_FUNCTION.as_ptr() as usize as u32) + 1,
-	];
-
-	let enabled = crate::pac::NVIC::is_enabled(crate::pac::Interrupt::SIO_IRQ_PROC0);
-	crate::pac::NVIC::mask(crate::pac::Interrupt::SIO_IRQ_PROC0);
-
-	'outer: loop {
-		for cmd in cmd_sequence.iter() {
-			debug!("Sending command {:x}...", *cmd);
-
-			// we drain before sending a 0
-			if *cmd == 0 {
-				debug!("Draining FIFO...");
-				fifo.drain();
-				// core 1 may be waiting for fifo space
-				cortex_m::asm::sev();
-			}
-			debug!("Pushing to FIFO...");
-			fifo.write_blocking(*cmd);
-
-			debug!("Getting response from FIFO...");
-			let response = loop {
-				if let Some(x) = fifo.read() {
-					break x;
-				} else {
-					debug!("ST is {:x}", fifo.status());
-				}
-			};
-
-			// move to next state on correct response otherwise start over
-			debug!("Got {:x}", response);
-			if *cmd != response {
-				continue 'outer;
-			}
-		}
-		break;
-	}
-
-	if enabled {
-		unsafe { crate::pac::NVIC::unmask(crate::pac::Interrupt::SIO_IRQ_PROC0) };
-	}
-
-	debug!("Waiting for Core 1 to start...");
-	while !CORE1_START_FLAG.load(Ordering::Relaxed) {
-		cortex_m::asm::nop();
-	}
-	debug!("Core 1 started!!");
 }
 
 /// Gets the current video mode
@@ -1216,8 +1496,6 @@ pub fn get_num_scan_lines() -> u16 {
 /// Only run this function on Core 1.
 #[link_section = ".data"]
 unsafe extern "C" fn core1_main() -> u32 {
-	CORE1_START_FLAG.store(true, Ordering::Relaxed);
-
 	let mut video = RenderEngine::new();
 
 	// The LED pin was called `_pico_led` over in the `Hardware::build`
@@ -1335,831 +1613,6 @@ unsafe fn PIO0_IRQ_1() {
 	dma.ch[TIMING_DMA_CHAN]
 		.ch_al3_read_addr_trig
 		.write(|w| w.bits(buffer as *const _ as usize as u32));
-}
-
-impl RenderEngine {
-	// Initialise the main-thread resources
-	pub fn new() -> RenderEngine {
-		RenderEngine {
-			frame_count: 0,
-			// Should match the default value of TIMING_BUFFER and VIDEO_MODE
-			current_video_mode: crate::common::video::Mode::new(
-				crate::common::video::Timing::T640x480,
-				crate::common::video::Format::Text8x16,
-			),
-			// Should match the mode above
-			num_text_cols: 80,
-			num_text_rows: 30,
-		}
-	}
-
-	/// Do the start-of-frame setup
-	fn frame_start(&mut self) {
-		trace!("Frame {}", self.frame_count);
-		self.frame_count += 1;
-
-		// Update video mode only on first line of video
-		if VIDEO_MODE.get_mode() != self.current_video_mode {
-			self.current_video_mode = VIDEO_MODE.get_mode();
-			match self.current_video_mode.timing() {
-				crate::common::video::Timing::T640x400 => unsafe {
-					TIMING_BUFFER = TimingBuffer::make_640x400();
-				},
-				crate::common::video::Timing::T640x480 => unsafe {
-					TIMING_BUFFER = TimingBuffer::make_640x480();
-				},
-				crate::common::video::Timing::T800x600 => {
-					panic!("Can't do 800x600");
-				}
-			}
-			self.num_text_cols = self.current_video_mode.text_width().unwrap_or(0) as usize;
-			self.num_text_rows = self.current_video_mode.text_height().unwrap_or(0) as usize;
-		}
-	}
-
-	/// Draw a line of pixels into the relevant pixel buffer (either
-	/// [`PIXEL_DATA_BUFFER_ODD`] or [`PIXEL_DATA_BUFFER_EVEN`]).
-	///
-	/// The `current_line_num` goes from `0..NUM_LINES`.
-	#[link_section = ".data"]
-	pub fn draw_next_line(&mut self, current_line_num: u16) {
-		// Pick a buffer to render into based on the line number we are drawing.
-		// It's safe to write to this buffer because it's the the other one that
-		// is currently being DMA'd out to the Pixel SM.
-		let scan_line_buffer = unsafe {
-			if (current_line_num & 1) == 0 {
-				&mut PIXEL_DATA_BUFFER_EVEN
-			} else {
-				&mut PIXEL_DATA_BUFFER_ODD
-			}
-		};
-
-		match self.current_video_mode.format() {
-			crate::common::video::Format::Text8x16 => {
-				self.draw_next_line_text::<16>(&font16::FONT, scan_line_buffer, current_line_num)
-			}
-			crate::common::video::Format::Text8x8 => {
-				self.draw_next_line_text::<8>(&font8::FONT, scan_line_buffer, current_line_num)
-			}
-			_ => {
-				// Draw nothing
-			}
-		};
-	}
-
-	/// Draw a line of pixels into the relevant pixel buffer (either
-	/// [`PIXEL_DATA_BUFFER_ODD`] or [`PIXEL_DATA_BUFFER_EVEN`]) using the 8x16
-	/// font.
-	///
-	/// The `current_line_num` goes from `0..NUM_LINES`.
-	#[link_section = ".data"]
-	pub fn draw_next_line_text<const GLYPH_HEIGHT: usize>(
-		&mut self,
-		font: &Font,
-		scan_line_buffer: &mut LineBuffer,
-		current_line_num: u16,
-	) {
-		// Convert our position in scan-lines to a text row, and a line within each glyph on that row
-		let text_row = current_line_num as usize / GLYPH_HEIGHT;
-		let font_row = current_line_num as usize % GLYPH_HEIGHT;
-
-		if text_row < self.num_text_rows {
-			// Note (unsafe): We could stash the char array inside `self`
-			// but at some point we are going to need one CPU rendering
-			// the text, and the other CPU running code and writing to
-			// the buffer. This might be Undefined Behaviour, but
-			// unfortunately real-time video is all about shared mutable
-			// state. At least our platform is fixed, so we can simply
-			// test if it works, for some given version of the Rust compiler.
-			let row_slice = unsafe {
-				&GLYPH_ATTR_ARRAY
-					[(text_row * self.num_text_cols)..((text_row + 1) * self.num_text_cols)]
-			};
-			// Every font look-up we are about to do for this row will
-			// involve offsetting by the row within each glyph. As this
-			// is the same for every glyph on this row, we calculate a
-			// new pointer once, in advance, and save ourselves an
-			// addition each time around the loop.
-			let font_ptr = unsafe { font.data.as_ptr().add(font_row) };
-
-			// Get a pointer into our scan-line buffer
-			let mut scan_line_buffer_ptr = scan_line_buffer.pixels.as_mut_ptr();
-
-			// Convert from characters to coloured pixels, using the font as a look-up table.
-
-			for glyphattr in row_slice.iter() {
-				let index = (glyphattr.glyph().0 as usize) * GLYPH_HEIGHT;
-				let attr = glyphattr.attr();
-
-				unsafe {
-					// Note (unsafe): We use pointer arithmetic here because we
-					// can't afford a bounds-check on an array. This is safe
-					// because the font is `256 * width` bytes long and we can't
-					// index more than `255 * width` bytes into it.
-					let mono_pixels = *font_ptr.add(index);
-
-					// 0bXX------
-					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels >> 6);
-					scan_line_buffer_ptr.write(pair);
-					// 0b--XX----
-					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels >> 4);
-					scan_line_buffer_ptr.offset(1).write(pair);
-					// 0b----XX--
-					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels >> 2);
-					scan_line_buffer_ptr.offset(2).write(pair);
-					// 0b------XX
-					let pair = TEXT_COLOUR_LOOKUP.lookup(attr, mono_pixels);
-					scan_line_buffer_ptr.offset(3).write(pair);
-
-					scan_line_buffer_ptr = scan_line_buffer_ptr.offset(4);
-				}
-			}
-		}
-	}
-}
-
-impl Default for RenderEngine {
-	fn default() -> Self {
-		RenderEngine::new()
-	}
-}
-
-impl TextConsole {
-	/// Create a TextConsole.
-	///
-	/// Has no buffer associated with it
-	pub const fn new() -> TextConsole {
-		TextConsole {
-			current_row: AtomicU16::new(0),
-			current_col: AtomicU16::new(0),
-			text_buffer: AtomicPtr::new(core::ptr::null_mut()),
-			// White on Black, with the default palette
-			current_attr: AtomicU8::new(
-				Attr::new(
-					TextForegroundColour::WHITE,
-					TextBackgroundColour::BLACK,
-					false,
-				)
-				.0,
-			),
-		}
-	}
-
-	/// Update the text buffer we are using.
-	///
-	/// Will reset the cursor. The screen is not cleared.
-	pub fn set_text_buffer(
-		&self,
-		text_buffer: &'static mut [GlyphAttr; MAX_TEXT_ROWS * MAX_TEXT_COLS],
-	) {
-		self.text_buffer
-			.store(text_buffer.as_mut_ptr(), Ordering::Relaxed)
-	}
-
-	/// Place a single Code Page 850 encoded 8-bit character on the screen.
-	///
-	/// Adjusts the current row and column automatically. Also understands
-	/// Carriage Return and New Line bytes.
-	pub fn write_font_glyph(&self, glyph: Glyph) {
-		// Load from global state
-		let mut row = self.current_row.load(Ordering::Relaxed);
-		let mut col = self.current_col.load(Ordering::Relaxed);
-		let buffer = self.text_buffer.load(Ordering::Relaxed);
-
-		if !buffer.is_null() {
-			self.write_at(glyph, buffer, &mut row, &mut col);
-			// Push back to global state
-			self.current_row.store(row, Ordering::Relaxed);
-			self.current_col.store(col, Ordering::Relaxed);
-		}
-	}
-
-	/// Moves the text cursor to the specified row and column.
-	///
-	/// If a value is out of bounds, the cursor is not moved in that axis.
-	pub fn move_to(&self, row: u16, col: u16) {
-		let mode = VIDEO_MODE.get_mode();
-		let num_rows = mode.text_height().unwrap_or(0);
-		let num_cols = mode.text_width().unwrap_or(0);
-		if row < num_rows {
-			self.current_row.store(row, Ordering::Relaxed);
-		}
-		if col < num_cols {
-			self.current_col.store(col, Ordering::Relaxed);
-		}
-	}
-
-	/// Convert a Unicode Scalar Value to a font glyph.
-	///
-	/// Zero-width and modifier Unicode Scalar Values (e.g. `U+0301 COMBINING,
-	/// ACCENT`) are not supported. Normalise your Unicode before calling
-	/// this function.
-	fn map_char_to_glyph(input: char) -> Glyph {
-		// This fixed table only works for the default font. When we support
-		// changing font, we will need to plug-in a different table for each font.
-		let index = match input {
-			'\u{0000}'..='\u{007F}' => input as u8,
-			'\u{00A0}' => 255, // NBSP
-			'\u{00A1}' => 173, // ¡
-			'\u{00A2}' => 189, // ¢
-			'\u{00A3}' => 156, // £
-			'\u{00A4}' => 207, // ¤
-			'\u{00A5}' => 190, // ¥
-			'\u{00A6}' => 221, // ¦
-			'\u{00A7}' => 245, // §
-			'\u{00A8}' => 249, // ¨
-			'\u{00A9}' => 184, // ©
-			'\u{00AA}' => 166, // ª
-			'\u{00AB}' => 174, // «
-			'\u{00AC}' => 170, // ¬
-			'\u{00AD}' => 240, // SHY
-			'\u{00AE}' => 169, // ®
-			'\u{00AF}' => 238, // ¯
-			'\u{00B0}' => 248, // °
-			'\u{00B1}' => 241, // ±
-			'\u{00B2}' => 253, // ²
-			'\u{00B3}' => 252, // ³
-			'\u{00B4}' => 239, // ´
-			'\u{00B5}' => 230, // µ
-			'\u{00B6}' => 244, // ¶
-			'\u{00B7}' => 250, // ·
-			'\u{00B8}' => 247, // ¸
-			'\u{00B9}' => 251, // ¹
-			'\u{00BA}' => 167, // º
-			'\u{00BB}' => 175, // »
-			'\u{00BC}' => 172, // ¼
-			'\u{00BD}' => 171, // ½
-			'\u{00BE}' => 243, // ¾
-			'\u{00BF}' => 168, // ¿
-			'\u{00C0}' => 183, // À
-			'\u{00C1}' => 181, // Á
-			'\u{00C2}' => 182, // Â
-			'\u{00C3}' => 199, // Ã
-			'\u{00C4}' => 142, // Ä
-			'\u{00C5}' => 143, // Å
-			'\u{00C6}' => 146, // Æ
-			'\u{00C7}' => 128, // Ç
-			'\u{00C8}' => 212, // È
-			'\u{00C9}' => 144, // É
-			'\u{00CA}' => 210, // Ê
-			'\u{00CB}' => 211, // Ë
-			'\u{00CC}' => 222, // Ì
-			'\u{00CD}' => 214, // Í
-			'\u{00CE}' => 215, // Î
-			'\u{00CF}' => 216, // Ï
-			'\u{00D0}' => 209, // Ð
-			'\u{00D1}' => 165, // Ñ
-			'\u{00D2}' => 227, // Ò
-			'\u{00D3}' => 224, // Ó
-			'\u{00D4}' => 226, // Ô
-			'\u{00D5}' => 229, // Õ
-			'\u{00D6}' => 153, // Ö
-			'\u{00D7}' => 158, // ×
-			'\u{00D8}' => 157, // Ø
-			'\u{00D9}' => 235, // Ù
-			'\u{00DA}' => 233, // Ú
-			'\u{00DB}' => 234, // Û
-			'\u{00DC}' => 154, // Ü
-			'\u{00DD}' => 237, // Ý
-			'\u{00DE}' => 232, // Þ
-			'\u{00DF}' => 225, // ß
-			'\u{00E0}' => 133, // à
-			'\u{00E1}' => 160, // á
-			'\u{00E2}' => 131, // â
-			'\u{00E3}' => 198, // ã
-			'\u{00E4}' => 132, // ä
-			'\u{00E5}' => 134, // å
-			'\u{00E6}' => 145, // æ
-			'\u{00E7}' => 135, // ç
-			'\u{00E8}' => 138, // è
-			'\u{00E9}' => 130, // é
-			'\u{00EA}' => 136, // ê
-			'\u{00EB}' => 137, // ë
-			'\u{00EC}' => 141, // ì
-			'\u{00ED}' => 161, // í
-			'\u{00EE}' => 140, // î
-			'\u{00EF}' => 139, // ï
-			'\u{00F0}' => 208, // ð
-			'\u{00F1}' => 164, // ñ
-			'\u{00F2}' => 149, // ò
-			'\u{00F3}' => 162, // ó
-			'\u{00F4}' => 147, // ô
-			'\u{00F5}' => 228, // õ
-			'\u{00F6}' => 148, // ö
-			'\u{00F7}' => 246, // ÷
-			'\u{00F8}' => 155, // ø
-			'\u{00F9}' => 151, // ù
-			'\u{00FA}' => 163, // ú
-			'\u{00FB}' => 150, // û
-			'\u{00FC}' => 129, // ü
-			'\u{00FD}' => 236, // ý
-			'\u{00FE}' => 231, // þ
-			'\u{00FF}' => 152, // ÿ
-			'\u{0131}' => 213, // ı
-			'\u{0192}' => 159, // ƒ
-			'\u{2017}' => 242, // ‗
-			'\u{2500}' => 196, // ─
-			'\u{2502}' => 179, // │
-			'\u{250C}' => 218, // ┌
-			'\u{2510}' => 191, // ┐
-			'\u{2514}' => 192, // └
-			'\u{2518}' => 217, // ┘
-			'\u{251C}' => 195, // ├
-			'\u{2524}' => 180, // ┤
-			'\u{252C}' => 194, // ┬
-			'\u{2534}' => 193, // ┴
-			'\u{253C}' => 197, // ┼
-			'\u{2550}' => 205, // ═
-			'\u{2551}' => 186, // ║
-			'\u{2554}' => 201, // ╔
-			'\u{2557}' => 187, // ╗
-			'\u{255A}' => 200, // ╚
-			'\u{255D}' => 188, // ╝
-			'\u{2560}' => 204, // ╠
-			'\u{2563}' => 185, // ╣
-			'\u{2566}' => 203, // ╦
-			'\u{2569}' => 202, // ╩
-			'\u{256C}' => 206, // ╬
-			'\u{2580}' => 223, // ▀
-			'\u{2584}' => 220, // ▄
-			'\u{2588}' => 219, // █
-			'\u{2591}' => 176, // ░
-			'\u{2592}' => 177, // ▒
-			'\u{2593}' => 178, // ▓
-			'\u{25A0}' => 254, // ■
-			_ => b'?',
-		};
-		Glyph(index)
-	}
-
-	/// Put a single character at a specified point on screen.
-	///
-	/// The character is relative to the current font.
-	fn write_at(&self, glyph: Glyph, buffer: *mut GlyphAttr, row: &mut u16, col: &mut u16) {
-		let mode = VIDEO_MODE.get_mode();
-		let num_rows = mode.text_height().unwrap_or(0) as usize;
-		let num_cols = mode.text_width().unwrap_or(0) as usize;
-		let attr = Attr(self.current_attr.load(Ordering::Relaxed));
-
-		if glyph.0 == b'\r' {
-			*col = 0;
-		} else if glyph.0 == b'\n' {
-			*col = 0;
-			*row += 1;
-		} else {
-			let offset = (*col as usize) + (num_cols * (*row as usize));
-			// Note (safety): This is safe as we bound `col` and `row`
-			unsafe {
-				buffer
-					.add(offset)
-					.write_volatile(GlyphAttr::new(glyph, attr))
-			};
-			*col += 1;
-		}
-		if *col == (num_cols as u16) {
-			*col = 0;
-			*row += 1;
-		}
-
-		if *row == (num_rows as u16) {
-			// Stay on last line
-			*row = (num_rows - 1) as u16;
-
-			unsafe { core::ptr::copy(buffer.add(num_cols), buffer, num_cols * (num_rows - 1)) };
-
-			for blank_col in 0..num_cols {
-				let offset = blank_col + (num_cols * (*row as usize));
-				unsafe {
-					buffer
-						.add(offset)
-						.write_volatile(GlyphAttr::new(Glyph(b' '), attr))
-				};
-			}
-		}
-	}
-
-	pub fn change_attr(&self, attr: Attr) {
-		let value = attr.0;
-		self.current_attr.store(value, Ordering::Relaxed);
-	}
-}
-
-unsafe impl Sync for TextConsole {}
-
-impl core::fmt::Write for &TextConsole {
-	/// Allows us to call `writeln!(some_text_console, "hello")`
-	fn write_str(&mut self, s: &str) -> core::fmt::Result {
-		// Load from global state
-		let mut row = self.current_row.load(Ordering::Relaxed);
-		let mut col = self.current_col.load(Ordering::Relaxed);
-		let buffer = self.text_buffer.load(Ordering::Relaxed);
-
-		if !buffer.is_null() {
-			for ch in s.chars() {
-				let b = TextConsole::map_char_to_glyph(ch);
-				self.write_at(b, buffer, &mut row, &mut col);
-			}
-
-			// Push back to global state
-			self.current_row.store(row, Ordering::Relaxed);
-			self.current_col.store(col, Ordering::Relaxed);
-		}
-
-		Ok(())
-	}
-}
-
-impl LineBuffer {
-	/// Convert the line buffer to a 32-bit address that the DMA engine understands.
-	fn as_ptr(&self) -> u32 {
-		self as *const _ as usize as u32
-	}
-}
-
-impl SyncPolarity {
-	const fn enabled(&self) -> bool {
-		match self {
-			SyncPolarity::Positive => true,
-			SyncPolarity::Negative => false,
-		}
-	}
-
-	const fn disabled(&self) -> bool {
-		match self {
-			SyncPolarity::Positive => false,
-			SyncPolarity::Negative => true,
-		}
-	}
-}
-
-impl ScanlineTimingBuffer {
-	const CLOCKS_PER_PIXEL: u32 = 6;
-
-	/// Create a timing buffer for each scan-line in the V-Sync visible portion.
-	///
-	/// The timings are in the order (front-porch, sync, back-porch, visible) and are in pixel clocks.
-	const fn new_v_visible(
-		hsync: SyncPolarity,
-		vsync: SyncPolarity,
-		timings: (u32, u32, u32, u32),
-	) -> ScanlineTimingBuffer {
-		ScanlineTimingBuffer {
-			data: [
-				// Front porch (as per the spec)
-				Self::make_timing(
-					timings.0 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.disabled(),
-					Some(1),
-				),
-				// Sync pulse (as per the spec)
-				Self::make_timing(
-					timings.1 * Self::CLOCKS_PER_PIXEL,
-					hsync.enabled(),
-					vsync.disabled(),
-					None,
-				),
-				// Back porch. Adjusted by a few clocks to account for interrupt +
-				// PIO SM start latency.
-				Self::make_timing(
-					(timings.2 * Self::CLOCKS_PER_PIXEL) - 5,
-					hsync.disabled(),
-					vsync.disabled(),
-					None,
-				),
-				// Visible portion. It also triggers the IRQ to start pixels
-				// moving. Adjusted to compensate for changes made to previous
-				// period to ensure scan-line remains at correct length.
-				Self::make_timing(
-					(timings.3 * Self::CLOCKS_PER_PIXEL) + 5,
-					hsync.disabled(),
-					vsync.disabled(),
-					Some(0),
-				),
-			],
-		}
-	}
-
-	/// Create a timing buffer for each scan-line in the V-Sync front-porch and back-porch
-	const fn new_v_porch(
-		hsync: SyncPolarity,
-		vsync: SyncPolarity,
-		timings: (u32, u32, u32, u32),
-	) -> ScanlineTimingBuffer {
-		ScanlineTimingBuffer {
-			data: [
-				// Front porch (as per the spec)
-				Self::make_timing(
-					timings.0 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.disabled(),
-					Some(1),
-				),
-				// Sync pulse (as per the spec)
-				Self::make_timing(
-					timings.1 * Self::CLOCKS_PER_PIXEL,
-					hsync.enabled(),
-					vsync.disabled(),
-					None,
-				),
-				// Back porch.
-				Self::make_timing(
-					timings.2 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.disabled(),
-					None,
-				),
-				// Visible portion.
-				Self::make_timing(
-					timings.3 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.disabled(),
-					None,
-				),
-			],
-		}
-	}
-
-	/// Create a timing buffer for each scan-line in the V-Sync pulse
-	const fn new_v_pulse(
-		hsync: SyncPolarity,
-		vsync: SyncPolarity,
-		timings: (u32, u32, u32, u32),
-	) -> ScanlineTimingBuffer {
-		ScanlineTimingBuffer {
-			data: [
-				// Front porch (as per the spec)
-				Self::make_timing(
-					timings.0 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.enabled(),
-					Some(1),
-				),
-				// Sync pulse (as per the spec)
-				Self::make_timing(
-					timings.1 * Self::CLOCKS_PER_PIXEL,
-					hsync.enabled(),
-					vsync.enabled(),
-					None,
-				),
-				// Back porch.
-				Self::make_timing(
-					timings.2 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.enabled(),
-					None,
-				),
-				// Visible portion.
-				Self::make_timing(
-					timings.3 * Self::CLOCKS_PER_PIXEL,
-					hsync.disabled(),
-					vsync.enabled(),
-					None,
-				),
-			],
-		}
-	}
-
-	/// Generate a 32-bit value we can send to the Timing FIFO.
-	///
-	/// * `period` - The length of this portion of the scan-line, in system clock ticks
-	/// * `hsync` - true if the H-Sync pin should be high during this period, else false
-	/// * `vsync` - true if the H-Sync pin should be high during this period, else false
-	/// * `raise_irq` - true the timing statemachine should raise an IRQ at the start of this period
-	///
-	/// Returns a 32-bit value you can post to the Timing FIFO.
-	const fn make_timing(period: u32, hsync: bool, vsync: bool, raise_irq: Option<u8>) -> u32 {
-		let command = if let Some(irq_index) = raise_irq {
-			// This command sets IRQ 0
-			pio::InstructionOperands::IRQ {
-				clear: false,
-				wait: false,
-				index: irq_index,
-				relative: false,
-			}
-			.encode()
-		} else {
-			// This command is a no-op (it moves Y into Y)
-			pio::InstructionOperands::MOV {
-				destination: pio::MovDestination::Y,
-				op: pio::MovOperation::None,
-				source: pio::MovSource::Y,
-			}
-			.encode()
-		} as u32;
-		let mut value: u32 = 0;
-		if hsync {
-			value |= 1 << 0;
-		}
-		if vsync {
-			value |= 1 << 1;
-		}
-		value |= (period - 6) << 2;
-		value | command << 16
-	}
-}
-
-impl TimingBuffer {
-	/// Make a timing buffer suitable for 640 x 400 @ 70 Hz
-	pub const fn make_640x400() -> TimingBuffer {
-		TimingBuffer {
-			visible_line: ScanlineTimingBuffer::new_v_visible(
-				SyncPolarity::Negative,
-				SyncPolarity::Positive,
-				(16, 96, 48, 640),
-			),
-			vblank_porch_buffer: ScanlineTimingBuffer::new_v_porch(
-				SyncPolarity::Negative,
-				SyncPolarity::Positive,
-				(16, 96, 48, 640),
-			),
-			vblank_sync_buffer: ScanlineTimingBuffer::new_v_pulse(
-				SyncPolarity::Negative,
-				SyncPolarity::Positive,
-				(16, 96, 48, 640),
-			),
-			visible_lines_ends_at: 399,
-			front_porch_end_at: 399 + 12,
-			sync_pulse_ends_at: 399 + 12 + 2,
-			back_porch_ends_at: 399 + 12 + 2 + 35,
-		}
-	}
-
-	/// Make a timing buffer suitable for 640 x 480 @ 60 Hz
-	pub const fn make_640x480() -> TimingBuffer {
-		TimingBuffer {
-			visible_line: ScanlineTimingBuffer::new_v_visible(
-				SyncPolarity::Negative,
-				SyncPolarity::Negative,
-				(16, 96, 48, 640),
-			),
-			vblank_porch_buffer: ScanlineTimingBuffer::new_v_porch(
-				SyncPolarity::Negative,
-				SyncPolarity::Negative,
-				(16, 96, 48, 640),
-			),
-			vblank_sync_buffer: ScanlineTimingBuffer::new_v_pulse(
-				SyncPolarity::Negative,
-				SyncPolarity::Negative,
-				(16, 96, 48, 640),
-			),
-			visible_lines_ends_at: 479,
-			front_porch_end_at: 479 + 10,
-			sync_pulse_ends_at: 479 + 10 + 2,
-			back_porch_ends_at: 479 + 10 + 2 + 33,
-		}
-	}
-}
-
-impl RGBColour {
-	/// Make an [`RGBColour`] from a 24-bit RGB triplet.
-	///
-	/// Only the top 4 bits of each colour channel are retained, as RGB colour
-	/// is a 12-bit value.
-	pub const fn from_24bit(red: u8, green: u8, blue: u8) -> RGBColour {
-		let red4: u16 = ((red >> 4) & 0x00F) as u16;
-		let green4: u16 = ((green >> 4) & 0x00F) as u16;
-		let blue4: u16 = ((blue >> 4) & 0x00F) as u16;
-		RGBColour((blue4 << 8) | (green4 << 4) | red4)
-	}
-
-	/// Make an [`RGBColour`] from a 12-bit RGB triplet.
-	///
-	/// Only the bottom 4 bits of each colour channel are retained, as RGB colour
-	/// is a 12-bit value.
-	pub const fn from_12bit(red: u8, green: u8, blue: u8) -> RGBColour {
-		let red4: u16 = (red & 0x00F) as u16;
-		let green4: u16 = (green & 0x00F) as u16;
-		let blue4: u16 = (blue & 0x00F) as u16;
-		RGBColour((blue4 << 8) | (green4 << 4) | red4)
-	}
-
-	/// Get the red component as an 8-bit value
-	pub const fn red8(self) -> u8 {
-		let red4 = self.0 & 0x0F;
-		(red4 << 4) as u8
-	}
-
-	/// Get the green component as an 8-bit value
-	pub const fn green8(self) -> u8 {
-		let green4 = (self.0 >> 4) & 0x0F;
-		(green4 << 4) as u8
-	}
-
-	/// Get the blue component as an 8-bit value
-	pub const fn blue8(self) -> u8 {
-		let blue4 = (self.0 >> 8) & 0x0F;
-		(blue4 << 4) as u8
-	}
-}
-
-impl From<RGBColour> for crate::common::video::RGBColour {
-	fn from(val: RGBColour) -> crate::common::video::RGBColour {
-		let red = val.red8();
-		let green = val.green8();
-		let blue = val.blue8();
-		crate::common::video::RGBColour::from_rgb(red, green, blue)
-	}
-}
-
-impl From<crate::common::video::RGBColour> for RGBColour {
-	fn from(val: crate::common::video::RGBColour) -> RGBColour {
-		let red = val.red();
-		let green = val.green();
-		let blue = val.blue();
-		RGBColour::from_24bit(red, green, blue)
-	}
-}
-
-impl RGBPair {
-	pub const fn from_pixels(first: RGBColour, second: RGBColour) -> RGBPair {
-		let first: u32 = first.0 as u32;
-		let second: u32 = second.0 as u32;
-		RGBPair((second << 16) | first)
-	}
-}
-
-impl GlyphAttr {
-	/// Make a new glyph/attribute pair.
-	pub const fn new(glyph: Glyph, attr: Attr) -> GlyphAttr {
-		let value: u16 = (glyph.0 as u16) + ((attr.0 as u16) << 8);
-		GlyphAttr(value)
-	}
-
-	/// Get the glyph component of this pair.
-	pub const fn glyph(self) -> Glyph {
-		Glyph(self.0 as u8)
-	}
-
-	/// Get the attribute component of this pair.
-	pub const fn attr(self) -> Attr {
-		Attr((self.0 >> 8) as u8)
-	}
-}
-
-impl AtomicModeWrapper {
-	/// Construct a new [`AtomicModeWrapper`] with the given value.
-	const fn new(mode: crate::common::video::Mode) -> AtomicModeWrapper {
-		AtomicModeWrapper {
-			value: AtomicU8::new(mode.as_u8()),
-		}
-	}
-
-	/// Set a new video mode.
-	fn set_mode(&self, mode: crate::common::video::Mode) {
-		self.value.store(mode.as_u8(), Ordering::SeqCst);
-	}
-
-	/// Get the current video mode.
-	fn get_mode(&self) -> crate::common::video::Mode {
-		let value = self.value.load(Ordering::SeqCst);
-		// Safety: the 'set_mode' function ensure this is always valid.
-		unsafe { crate::common::video::Mode::from_u8(value) }
-	}
-}
-
-impl TextColourLookup {
-	const fn blank() -> TextColourLookup {
-		TextColourLookup {
-			entries: [RGBPair(0); 512],
-		}
-	}
-
-	fn init(&mut self, palette: &[AtomicU16]) {
-		for (fg, fg_colour) in palette.iter().take(16).enumerate() {
-			for (bg, bg_colour) in palette.iter().take(8).enumerate() {
-				let attr = unsafe {
-					Attr::new(
-						TextForegroundColour::new_unchecked(fg as u8),
-						TextBackgroundColour::new_unchecked(bg as u8),
-						false,
-					)
-				};
-				for pixels in 0..=3 {
-					let index: usize = (((attr.0 & 0x7F) as usize) << 2) | (pixels & 0x03) as usize;
-					let pair = RGBPair::from_pixels(
-						if pixels & 0x02 == 0x02 {
-							RGBColour(fg_colour.load(Ordering::Relaxed))
-						} else {
-							RGBColour(bg_colour.load(Ordering::Relaxed))
-						},
-						if pixels & 0x01 == 0x01 {
-							RGBColour(fg_colour.load(Ordering::Relaxed))
-						} else {
-							RGBColour(bg_colour.load(Ordering::Relaxed))
-						},
-					);
-					self.entries[index] = pair;
-				}
-			}
-		}
-	}
-
-	#[inline]
-	fn lookup(&self, attr: Attr, pixels: u8) -> RGBPair {
-		let index: usize = (((attr.0 & 0x7F) as usize) << 2) | (pixels & 0x03) as usize;
-		unsafe { core::ptr::read(self.entries.as_ptr().add(index)) }
-	}
 }
 
 // -----------------------------------------------------------------------------
